Finished restructuring / refactoring / renaming work, albeit for any pesky errors which may pop up.
