+++ code/evaluation/README.md
@@ -1,0 +1,4 @@
+# Evaluation scripts
+
+This subdirectory contains scripts which allow the user to evaluate the accuracy of various trained spiking neural network models on a held-out test dataset. Most of these
+scripts are no longer useful, as the evaluation may have been built into the "test" phase of the scripts found in the "train" folder.
