+++ code/plotting/README.md
@@ -1,0 +1,3 @@
+# Plotting scripts
+
+This subdirectory contains scripts which allow the user to create plots of the performance of various neural network models over the course of their training phase.
