+++ code/plotting/plot_all_csnn_performance.py
