+++ code/train/csnn_mnist_brian2.py
