+++ code/train/csnn_pc_weight_habituation_mnist.py
@@ -21,7 +21,7 @@ b.log_level_error()
 top_level_path = '../../'
 MNIST_data_path = top_level_path + 'data/'
 results_path = top_level_path + 'results/'
-model_name = 'scnn_pc_weight_habituation'
+model_name = 'csnn_pc_weight_habituation'
 
 performance_dir = top_level_path + 'performance/' + model_name + '/'
 activity_dir = top_level_path + 'activity/' + model_name + '/'
