+++ code/visualization/csnn_weight_visualization.py
